requirements.txt
flask
gunicorn
python-dotenv
flask-wtf
pymongo
passlib
WTForms
idna
requests
pip
attrs
argparse
toml
wheel
py
lxml
Pillow
Jinja2
MarkupSafe
Werkzeug
click
itsdangerous
certifi
setuptools
